Start to reorganize tests
This commit is the start of a refactoring which will move from each code
generator having one monolithic test into many separate sub-tests. This
should enable smaller and more focused tests which benefits bringing new
bindings generators online (can get some tests working before others)
and also allows writing tests for some generators but not others.

The way that this is currently architected is that there's a new
`tests/runtime/*` directory in which a test-per-directory lives. For
each tests `tests/runtime/foo` there will be a few conventions:

* `tests/runtime/foo/imports.witx` - these are what hosts must
  implement and what the wasm can import from the host. This file is
  required to exist.
* `tests/runtime/foo/exports.witx` - this is what wasms must implement
  and what hosts can call. This file must exist.
* `tests/runtime/foo/host.*` - these are separate host implementations
  to execute the wasm code. The extension of the file indicates which
  generator it's used for (e.g. right now `host.{ts,rs,py}` are
  supported). If `host.py` doesn't exist for the test `foo` then testing
  the Python code generator will skip the test `foo`. Otherwise it's
  expected to work. At least one `host.*` file is required to work, but
  there's no need to have a file-per-host.
* `tests/runtime/foo/wasm.*` - these are separate guest compiled-to-wasm
  implementations. Currently only `wasm.{c,rs}` are supported, but the
  intent is to support `wasm.js` for #71 as well. Not all languages need
  to have an implementation here either, they'll be skipped if omitted.

The general goal is that adding a test for some witx functionality
should be "drop a few files in `tests/runtime/your_new_test` and you're
good to go. Over time more generators can add more tests if it becomes
necessary, and generator-specific tests can also be added to this
directory such as the one test exercising JS instantiation added here
too.

Overall I think this turned out quite well. The biggest downside is that
Rust-compiled-to-wasm needs to list explicitly each test that's being
compiled to wasm, which is a bit of a bummer.