Bump version of dependency crates (#438)
* Bump version of dependency crates

This commit updates this repository to published versions of the
`wasm-tools` family of crates. Additionally Wasmtime is updated with the
same upstream versions of the `wasm-tools` crates. Put together this
updates everything to the most recent version of the component model
with optional URLs encoded and an updated version of Wasmtime.

Part of this transition is the removal of
`wit_component::ComponentInterfaces` which is plumbed through as part of
this commit.

* Update to wasmtime main

* Fix demo build

* Update wasm-tools binary

* Fix cli build

* Install wasm-tools from crates.io, not git