Update to Wasmtime 12
This commit updates the Wasmtime dependency to the, currently
unreleased, Wasmtime 12.0.0 branch. A number of changes happened to
component translation in Wasmtime 12 to account for resources which
required changes here as well. Additionally Wasmtime 12 disallows empty
types in components which WASI was previously using, so this
additionally updates the preview1 shims and test components.