resource support for `componentize-js`
This is to support resource handle lifting and lowering from the guest's
point-of-view.  Unlike with the other, structural data types, resource bindings
must be either host- or guest-specific (i.e. the bindings are not symmetric).

I'll follow this up with a `componentize-js` PR that includes a bunch of tests.

Note that I had to update several dependencies to pull in
https://github.com/bytecodealliance/wasm-tools/pull/1218, which addresses the
case of a WIT world that `use`s a resource from an interface while also
exporting the interface.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>