fix: Remove a useless `if` to `setShowSpinner(true)`.
The code says: if the spinner must be shown, then let's set the spinner
to be shown. This is redundant.

This patch removes the `if (showSpinner)` to keep only the
`setShowSpinner(true)`, as it will have the same behaviour. Even if the
intend was to call `setShowSpinner` when `showSpinner` was `false`, it
wasn't working before but will work now.