feat: Change what `--js` produce.
So, this patch changes what `--js` does. Instead of generating a `$init`
function, it now mimics `instantiate`. The original `instantiate`
function is renamed to `_instantiate`, and `--js` creates a new
`instantiate` that wraps `_instantiate` with custom `getCoreModule`
and `instantiateCore` arguments/functions. Instead of compiling and
instantiating a Wasm module, it simply calls the code generated by
`wasm2js` that are stored in the `asmInit` and pass imports.