fix: don't overallocate string buffer while encoding into utf8
Recent dlmalloc-rs started validating that deallocation size is same as
the size that was allocated.
https://github.com/alexcrichton/dlmalloc-rs/blob/0.2.6/src/dlmalloc.rs#L1187

Since canonical abi string does not have "capacity" concept separate
from "length", we don't have a way to tell how much memory was
allocated to the buffer owner. So, canonical abi always assumes that
"length" is the same as "capacity" and deallocates the buffer with the
"length" size.

This means that if we overallocate the buffer and return the buffer with
a length less than the allocated size, dlmalloc-rs validation will fail.

This commit fixes the issue by lopping off the extra allocated space by
realloc with smaller size.