fix: call realloc exactly once for string encoding
Previously, we would call realloc multiple times to grow the buffer and
shrink it back down to the correct size.
However, realloc provided by wasi-preview1-component-adapter is quite
restrictive:
* It can't always be called multiple times during a single imported
  function call.
* It does not accept non-null old pointers.

This commit changes the code to call realloc exactly once, which is
compatible with those restrictions. The use of temporary buffers with
`TextEncoder.encode` is less efficient than the previous `encodeInto`,
but it allows us to keep the allocator in the adapter simple.
We can optimize it after we will be able to get rid of the adapter or
decide to complicate the allocator in the adapter.