fix: don't overallocate string buffer while encoding into utf8 (#444)
* fix: don't overallocate string buffer while encoding into utf8

Recent dlmalloc-rs started validating that deallocation size is same as
the size that was allocated.
https://github.com/alexcrichton/dlmalloc-rs/blob/0.2.6/src/dlmalloc.rs#L1187

Since canonical abi string does not have "capacity" concept separate
from "length", we don't have a way to tell how much memory was
allocated to the buffer owner. So, canonical abi always assumes that
"length" is the same as "capacity" and deallocates the buffer with the
"length" size.

This means that if we overallocate the buffer and return the buffer with
a length less than the allocated size, dlmalloc-rs validation will fail.

This commit fixes the issue by lopping off the extra allocated space by
realloc with smaller size.

* fix: call realloc exactly once for string encoding

Previously, we would call realloc multiple times to grow the buffer and
shrink it back down to the correct size.
However, realloc provided by wasi-preview1-component-adapter is quite
restrictive:
* It can't always be called multiple times during a single imported
  function call.
* It does not accept non-null old pointers.

This commit changes the code to call realloc exactly once, which is
compatible with those restrictions. The use of temporary buffers with
`TextEncoder.encode` is less efficient than the previous `encodeInto`,
but it allows us to keep the allocator in the adapter simple.
We can optimize it after we will be able to get rid of the adapter or
decide to complicate the allocator in the adapter.