fix: stdout/stderr pollable hang
This commit fixes an issue with pollables and
stdout/stderr. Platform-specific behavior on NodeJS requires special
treatment, as writes to stdout/stderr go across a port that is only
flushed when the worker *exits*.

To fix this issue, we forgo the ability to reject writes to
stdout/stderr, but update state immediately.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>