fix: allow objects through globalThis intrinsic (WebIDL) (#584)
This commit enables objects to be set on the global context and used
via the WebIDL integration in Jco. Previously, objects would be
detected as functions and then proxied into another function, which
prevented Object.create(O.prototype) calls from succeeding.

With this change, we add an additional check to allow classes to pass
through the proxying machinery of hte globalThis intrinsic, to avoid
having `O.prototype` (i.e. when used in the context of
`Object.create(O.prototype)`) evaluauate to `() => proxy(O.protoytpe)`

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>