feat(p2-shim): add easy to use p2 shim object
This commit introduces an import for `preview2-shim` that can be used
to more easily fulfill requirements for `instantiate` calls that are
exported from transpiled modules.

After creating a `WasiP2Shim` object, users should be able to call
`instantiate(undefined, shim.importObject())` and make use of all WASI
interfaces specified (or overriden) on the shim.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>