docs: fix rendering for | in code blocks in table (#626)
The `|` was not rendering correctly inside code blocks within tables when processed by `mdbook`.
https://github.com/rust-lang/mdBook/issues/637

Idealy, this should be addressed by the upstream. The root cause of this problem in `pulldown-cmark` has recently been fixed. https://github.com/pulldown-cmark/pulldown-cmark/issues/356
However, `mdbook` seems to expose the `pulldown-cmark` types as a public API, and is currently facing difficulties updating its dependencies.
https://github.com/rust-lang/mdBook/pull/2386
Therefore, the upstream fix may not available in any time soon.

As a workaround for the time being, use <code> tag instead of code blocks in tables with `|`.