refactor: split out jco transpile (#772)
* refactor(transpile): split out jco transpile functionality

This commit splits out functionality that powers `jco transpile` into
it's own package so it can be used independently of jco itself.

While this particular commit does not yet change `@bytecodealliance/jco` to rely on
`jco-transpile`, once a stable version of
`@bytecodealliance/jco-transpile` has been released, `jco` can move to
using this library directly in a manner similar to `componentize-js`.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

* fix(jco): features option in type generation, backwards compat

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>

---------

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>