chore(ci): remove pr title linting
PR title linting doesn't work very well, given that we do not want to
actually use `pull_request_target` due to security concerns.

Instead, it probably makes more sense for Jco to move to fast-forward
rebase merging only and keep the commit-based checks in place.

This means that we'll have to be a bit more rigorous around commits
that are merged into Jco (ensuring they represent the contents and are
significante enough to stand alone).