feat(std): add jco-std along with hono adapter
This commit adds the jco-std project, which contains helpers and
utilities for building JS WebAssembly components.

In particular, this commit contains a Hono adapter which enable Hono
to work with raw `wasi:http` *or* `fetch-event` APIs that are
supported by StarlingMonkey.

At present, `fetch-event` has a limitation in that the handler cannot
be run twice, so by default the `wasi:http` implementation is used for
Hono applications.