feat(p2-shim): accomodate versioned WASI import objects
This commit adds the typing to support versioned WASI import object
generation, dynamically appending the appropriate
version (e.g. `@0.2.3`) to import objects when `opts.asVersion` is
specified.

Generally 0.2.x versions are expected to be backwards compatible, but
it is sometimes useful to be able to control the strings in the object
that are generated.