fix(bindgen): declare within declare for reserved words
This commit fixes a bug during bindgen that caused `declare function`
to be used for invalid JS identifiers, but given that we are now using
modules (and not TS namespaces) this conflicts with the top level
module `declare` (e.g. `declare 'wasi:keyvalue...' {`).

This commit removes the `declare` keyword since it isn't necessary,
and the export renaming trick (`export { _keyword as keyword }`) still
works.

Resolves #624