refactor(jco): remove chalk dependency
This commit removes the `chalk` dependency. While we were not
vulnerable to the upstream malicious code injection that happened to
chalk and related deps, the chalk dependency is no longer needed when
using Node 20.x and above.

For node versions 20 or other platforms that do not have the
dependency, we use a polyfill that simply loses the colored output.

This is a breaking change in a sense as output will change and users
that are on node 18 and rely on the color output will get output that
is inconsistent from before, but removing this dependency is a high
priority at this point and the repercussions are likely minor from the
light semver breakage.

Link: https://github.com/chalk/chalk/issues/656