fix(jco): require utf8 decoder for flat string lifts
This commit fixes a bug by which strings loaded from memory during
async returns did not have the shared decoder (created via intrinsic)
present.

This commit stops short of reworking the dependency resolution code to
be closer to the site of usage (i.e. in the various `*Intrinsic` enum
members), but that can/should be done in the future.