fix(p2-shim): missing browser shim bits
This commit fixes an issue that was exposed when all imports are
checked for existence during import. The browser shims at present not
only are incomplete, but they're also missing key functions and
classes that should be present.

This commit unfortunately doesn't actually bring the browser
implementation up to snuff -- rather we simply add the placeholders
that should be there (and will fail if use is attempted at runtime).

Up until now imports were not checked for existence, but the
usefulness of explicit upfront errors (instead of less scruable ones
later) is a good tradeoff.

In the future an option to disable import checks may be added.